clean up atoi, update functions, improve header and Makefile

+++ Libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: dduffy-m <dduffy-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/16 16:12:28 by dduffy-m          #+#    #+#             */
-/*   Updated: 2023/11/16 16:15:25 by dduffy-m         ###   ########.fr       */
+/*   Updated: 2024/05/05 16:47:10 by dduffy-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
